Refactor Python Runtime Types
