+++ pystd/CMakeLists.txt
@@ -16,16 +16,16 @@ add_library(pystd STATIC
         src/builtin_runtime.cpp
         src/runtime_util.cpp
         src/function_runtime.cpp
-        src/pytypes/py_list.cpp
-        src/pytypes/py_set.cpp
-        src/pytypes/py_dict.cpp
-        src/pytypes/py_tuple.cpp
-        src/pytypes/py_str.cpp
-        src/pytypes/py_int.cpp
-        src/pytypes/py_float.cpp
-        src/pytypes/py_bool.cpp
+        src/pytypes/iterables/py_list.cpp
+        src/pytypes/iterables/py_set.cpp
+        src/pytypes/iterables/py_dict.cpp
+        src/pytypes/iterables/py_tuple.cpp
+        src/pytypes/primitives/py_str.cpp
+        src/pytypes/primitives/py_int.cpp
+        src/pytypes/primitives/py_float.cpp
+        src/pytypes/primitives/py_bool.cpp
         src/pytypes/py_object.cpp
-        src/pytypes/py_none.cpp
+        src/pytypes/primitives/py_none.cpp
         src/pytypes/py_function.cpp
         src/pytypes/py_method.cpp
         src/control_flow_runtime.cpp
