+++ pystd/src/pytypes/iterables/py_set.h
@@ -7,7 +7,7 @@
 #include <unordered_set>
 #include <utility>
 
-#include "pytypes/py_iter.h"
+#include "pytypes/iterables/py_iter.h"
 #include "pytypes/py_object.h"
